# Conflicts:
#	src/main/java/aplicacion/GestionUsuarios.java
#	src/main/java/baseDatos/DAOUsuarios.java

+++ src/main/java/aplicacion/Empresa.java
@@ -13,11 +13,11 @@ public class Empresa extends Usuario {
     private String CIF;
     private String direccion;
     private String telefono;
-    private Float saldo;
-    private Float saldobloqueado;
+    private float saldo;
+    private float saldobloqueado;
 
 
-    public Empresa(String idUsuario, String nombre, String CIF, Float saldo, Float saldobloqueado, String direccion, String telefono, boolean solicitadobaja, boolean autorizado) {
+    public Empresa(String idUsuario, String nombre, String CIF, float saldo, float saldobloqueado, String direccion, String telefono, boolean solicitadobaja, boolean autorizado) {
         super(idUsuario, solicitadobaja, autorizado);
         this.nombre = nombre;
         this.CIF = CIF;
@@ -63,19 +63,19 @@ public class Empresa extends Usuario {
         this.telefono = telefono;
     }
 
-    public Float getSaldo() {
+    public float getSaldo() {
         return saldo;
     }
 
-    public void setSaldo(Float saldo) {
+    public void setSaldo(int saldo) {
         this.saldo = saldo;
     }
 
-    public Float getSaldobloqueado() {
+    public float getSaldobloqueado() {
         return saldobloqueado;
     }
 
-    public void setSaldobloqueado(Float saldobloqueado) {
+    public void setSaldobloqueado(float saldobloqueado) {
         this.saldobloqueado = saldobloqueado;
     }
